Move term utils to separate file
