+++ src/store.js
@@ -20,7 +20,8 @@ import ClassOrder from './class-order'
 import { defaultGraphURI } from './data-factory-internal'
 import DataFactory from './data-factory'
 import Formula from './formula'
-import { ArrayIndexOf, isStatement, isStore, RDFArrayRemove } from './util'
+import { ArrayIndexOf, RDFArrayRemove } from './util'
+import { isStatement, isStore } from './utils/terms'
 import Statement from './statement'
 import Node from './node'
 import Variable from './variable'
