+++ src/util.js
@@ -30,25 +30,6 @@ export const appliedFactoryMethods = [
   'supports',
 ]
 
-export function isTerm(obj) {
-  return typeof obj === "object"
-    && obj !== null
-    && "termType" in obj
-    && "value" in obj
-}
-
-export function isStatement(obj) {
-  return typeof obj === "object" && obj !== null && "subject" in obj
-}
-
-export function isStore(obj) {
-  return typeof obj === "object" && obj !== null && "statements" in obj
-}
-
-export function isNamedNode(obj) {
-  return isTerm(obj) && obj.termType === "NamedNode"
-}
-
 /**
  * Loads ontologies of the data we load (this is the callback from the kb to
  * the fetcher).
